requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
email-validator==1.3.1
Flask-Login==0.6.3
